Whos_Rem/requirements.txt
arcade
python-vlc
PyAutoGUI